+++ packages/components/src/components/List/components/Header/components/SearchField/index.ts
@@ -1,0 +1,4 @@
+import { SearchField } from "./SearchField";
+
+export * from "./SearchField";
+export default SearchField;