+++ packages/components/src/components/List/components/ListStaticData.ts
@@ -1,7 +1,0 @@
-interface Props<T> {
-  data: T[];
-}
-
-export function ListStaticData<T = never>(ignoredProps: Props<T>) {
-  return null;
-}