+++ packages/components/src/components/List/model/loading/types.ts
@@ -1,6 +1,7 @@
 import type { PropertyRecord } from "@/components/List/model/types";
 import type { FilterMode } from "@/components/List/model/filter/types";
 import type { AsyncResource } from "@mittwald/react-use-promise";
+import type { SearchValue } from "@/components/List/model/search/types";
 
 type DataLoaderSortOptions<T> = PropertyRecord<T, "asc" | "desc">;
 
@@ -17,6 +18,7 @@ interface DataLoaderPaginationOptions {
 export interface DataLoaderOptions<T> {
   sorting?: DataLoaderSortOptions<T>;
   filtering?: DataLoaderFilterOptions<T>;
+  searchString?: SearchValue;
   pagination?: DataLoaderPaginationOptions;
 }
 
@@ -33,7 +35,7 @@ type AsyncResourceFactory<T> = (
   options?: DataLoaderOptions<T>,
 ) => AsyncResource<DataLoaderResult<T>>;
 
-type StaticData<T> = T[];
+type StaticData<T> = T[] | readonly T[];
 
 interface DynamicLoaderShape {
   manualSorting?: boolean;