+++ packages/components/src/components/Section/components/SectionHeader/SectionHeader.tsx
@@ -1,0 +1,58 @@
+import type { FC, PropsWithChildren } from "react";
+import React from "react";
+import styles from "./SectionHeader.module.scss";
+import clsx from "clsx";
+import type { PropsContext } from "@/lib/propsContext";
+import { PropsContextProvider } from "@/lib/propsContext";
+import type { PropsWithClassName } from "@/lib/types/props";
+import { TunnelExit, TunnelProvider } from "@mittwald/react-tunnel";
+
+export type SectionHeaderProps = PropsWithChildren & PropsWithClassName;
+
+export const SectionHeader: FC<SectionHeaderProps> = (props) => {
+  const { children, className } = props;
+
+  const rootClassName = clsx(styles.sectionHeader, className);
+
+  const propsContext: PropsContext = {
+    Switch: {
+      labelPosition: "leading",
+      tunnelId: "actions",
+    },
+    Button: {
+      size: "s",
+      tunnelId: "actions",
+    },
+    ContextMenuTrigger: {
+      tunnelId: "actions",
+      Button: {
+        tunnelId: null,
+      },
+    },
+    Action: {
+      tunnelId: "actions",
+      Button: {
+        tunnelId: null,
+      },
+    },
+    Heading: {
+      level: 2,
+      className: styles.heading,
+    },
+  };
+
+  return (
+    <header className={rootClassName}>
+      <PropsContextProvider props={propsContext}>
+        <TunnelProvider>
+          {children}
+          <div className={styles.actions}>
+            <TunnelExit id="actions" />
+          </div>
+        </TunnelProvider>
+      </PropsContextProvider>
+    </header>
+  );
+};
+
+export default SectionHeader;