+++ packages/components/src/components/propTypes/index.ts
@@ -40,6 +40,8 @@ import type { FlowComponentName } from "@/components/propTypes/types";
 import type { ContextualHelpProps } from "@/components/ContextualHelp";
 import type { PopoverProps } from "@/components/Popover";
 import type { ContextMenuSectionProps } from "@/components/ContextMenu/components/ContextMenuSection";
+import type { OverlayTriggerProps } from "@/components/OverlayTrigger";
+import type { ListProps } from "@/components/List";
 
 export * from "./types";
 
@@ -54,6 +56,7 @@ export interface FlowComponentPropsTypes {
   Content: ContentProps;
   ContextMenu: ContextMenuProps;
   ContextMenuSection: ContextMenuSectionProps;
+  ContextMenuTrigger: OverlayTriggerProps;
   ContextualHelp: ContextualHelpProps;
   CopyButton: CopyButtonProps;
   CounterBadge: CounterBadgeProps;
@@ -68,6 +71,7 @@ export interface FlowComponentPropsTypes {
   Label: LabelProps;
   LayoutCard: LayoutCardProps;
   Link: LinkProps;
+  List: ListProps;
   MenuItem: MenuItemProps;
   Modal: ModalProps;
   NumberField: NumberFieldProps;
@@ -101,6 +105,7 @@ const propsContextSupportingComponentsMap: Record<
   Content: true,
   ContextMenu: true,
   ContextMenuSection: true,
+  ContextMenuTrigger: true,
   ContextualHelp: true,
   CopyButton: true,
   CounterBadge: true,
@@ -115,6 +120,7 @@ const propsContextSupportingComponentsMap: Record<
   Label: true,
   LayoutCard: true,
   Link: true,
+  List: true,
   MenuItem: true,
   Modal: true,
   NumberField: true,