+++ packages/docs/src/content/03-components/content/labeled-value/examples/copy-button.tsx
@@ -2,9 +2,31 @@ import LabeledValue from "@mittwald/flow-react-components/LabeledValue";
 import Label from "@mittwald/flow-react-components/Label";
 import Content from "@mittwald/flow-react-components/Content";
 import CopyButton from "@mittwald/flow-react-components/CopyButton";
+import Link from "@mittwald/flow-react-components/Link";
+import Text from "@mittwald/flow-react-components/Text";
+import { IconExternalLink } from "@mittwald/flow-react-components/Icons";
 
-<LabeledValue>
-  <Label>Projekt ID</Label>
-  <Content>p-eWRhpRV</Content>
-  <CopyButton text="Projekt ID: p-eWRhpRV" />
-</LabeledValue>;
+<Row>
+  <LabeledValue>
+    <Label>Short-ID</Label>
+    <Content>p-wut3uw</Content>
+    <CopyButton text="p-wut3uw" />
+  </LabeledValue>
+
+  <LabeledValue>
+    <Label>Datenbankserver</Label>
+    <Content>
+      mysql-kq3v4g.ge-s-hopovh.db.project.host
+    </Content>
+    <CopyButton text="mysql-kq3v4g.ge-s-hopovh.db.project.host" />
+  </LabeledValue>
+
+  <LabeledValue>
+    <Label>Projektdomain</Label>
+    <Link>
+      <Text>p-lol3qe.project.space</Text>
+      <IconExternalLink />
+    </Link>
+    <CopyButton text="p-lol3qe.project.space" />
+  </LabeledValue>
+</Row>;